# docs/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  handler.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  handler.TokenResponse:
    properties:
      token:
        type: string
    type: object
...
  model.Admin:
    properties:
      id:
        type: string
      login:
        type: string
      password:
        type: string
    type: object
  model.AdminLoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  model.Brand:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
    type: object
  model.BrandInput:
    properties:
      name:
        type: string
    type: object
  model.BrandOutput:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.Buyer:
    properties:
      email:
        type: string
      favorites:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      id:
        type: string
      name:
        type: string
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      password:
        type: string
    type: object
  model.BuyerOutput:
    properties:
      email:
        type: string
      favorites:
        items:
          $ref: '#/definitions/model.ItemInfo'
        type: array
      id:
        type: string
      name:
        type: string
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
    type: object
...
  model.Category:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
    type: object
...
  model.Image:
    properties:
      id:
        type: integer
      item:
        $ref: '#/definitions/model.Item'
      item_id:
        type: integer
      url:
        type: string
    type: object
  model.Item:
    properties:
      article:
        type: string
      brand:
        $ref: '#/definitions/model.Brand'
      brand_id:
        type: integer
      category:
        $ref: '#/definitions/model.Category'
      category_id:
        type: integer
      description:
        type: string
      favoritedBy:
        items:
          $ref: '#/definitions/model.Buyer'
        type: array
      height:
        type: integer
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/model.Image'
        type: array
      length:
        type: integer
      material:
        $ref: '#/definitions/model.Material'
      material_id:
        type: integer
      name:
        type: string
      price:
        type: number
      price_with_discount:
        type: number
      quantity:
        type: integer
      seller:
        $ref: '#/definitions/model.Seller'
      seller_id:
        type: string
      weight:
        type: integer
      width:
        type: integer
    type: object
  model.ItemInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
  model.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.Material:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
    type: object
...
  model.Order:
    properties:
      buyer:
        $ref: '#/definitions/model.Buyer'
      buyer_id:
        type: string
      id:
        type: integer
      order_items:
        items:
          $ref: '#/definitions/model.OrderItem'
        type: array
      status:
        type: string
      total:
        type: number
    type: object
  model.OrderItem:
    properties:
      id:
        type: integer
      item:
        $ref: '#/definitions/model.Item'
      item_id:
        type: integer
      order:
        $ref: '#/definitions/model.Order'
      order_id:
        type: integer
      quantity:
        type: integer
      total:
        type: number
      unit_price:
        type: number
    type: object
  model.Seller:
    properties:
      balance:
        type: number
      email:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
      password:
        type: string
      shop_name:
        type: string
    type: object
info:
  contact: {}
paths:
  /admin/brand:
    delete:
      description: Deletes a brand by ID. Only accessible by admin.
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Brand ID
        in: query
        name: brand_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Brand deleted successfully
          schema:
            type: string
        "400":
          description: Invalid brand ID
          schema:
            type: string
        "401":
          description: You are not authorized to access this resource
          schema:
            type: string
        "500":
          description: Failed to delete brand
          schema:
            type: string
      summary: Delete a brand
      tags:
      - brands
    post:
      consumes:
      - application/json
      description: Creates a new brand with the provided details. Only accessible
        by admin.
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Brand data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.BrandInput'
      produces:
      - application/json
      responses:
        "201":
          description: Brand created successfully
          schema:
            type: string
        "400":
          description: Invalid input data
          schema:
            type: string
        "401":
          description: You are not authorized to access this resource
          schema:
            type: string
        "500":
          description: Failed to create brand
          schema:
            type: string
      summary: Create a new brand
      tags:
      - brands
...
  /admin/sign-up:
    post:
      consumes:
      - application/json
      description: Registers a new admin user
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Admin signup data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Admin'
      produces:
      - application/json
      responses:
        "201":
          description: Admin signed up successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to sign up
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Sign up as an admin
      tags:
      - admin
  /brand:
    get:
      description: Retrieves a list of all brands. Only accessible by admin.
      produces:
      - application/json
      responses:
        "200":
          description: List of brands
          schema:
            items:
              $ref: '#/definitions/model.BrandOutput'
            type: array
        "401":
          description: You are not authorized to access this resource
          schema:
            type: string
        "500":
          description: Failed to get brand list
          schema:
            type: string
      summary: Get list of brands
      tags:
      - brands
  /buyer:
    get:
      description: Retrieve buyer information by ID
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Buyer data
          schema:
            $ref: '#/definitions/model.BuyerOutput'
        "403":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Buyer not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get buyer information
      tags:
      - Buyers
    patch:
      consumes:
      - application/json
      description: Update buyer details
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Buyer data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Buyer'
      produces:
      - application/json
      responses:
        "200":
          description: Buyer updated successfully
          schema:
            type: string
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to update buyer
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update buyer information
      tags:
      - Buyers
...
  /sign-in/admin:
    post:
      consumes:
      - application/json
      description: Logs in an admin user
      parameters:
      - description: Admin login data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authentication token
          schema:
            $ref: '#/definitions/handler.TokenResponse'
        "400":
          description: Invalid login data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Sign in as an admin
      tags:
      - admin
  /sign-in/buyer:
    post:
      consumes:
      - application/json
      description: Authenticate a buyer with email and password
      parameters:
      - description: Login data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authenticated buyer data
          schema:
            $ref: '#/definitions/model.BuyerOutput'
        "400":
          description: Invalid login data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Buyer login
      tags:
      - Buyers
...
  /sign-up/buyer:
    post:
      consumes:
      - application/json
      description: Register a new buyer in the system
      parameters:
      - description: Buyer data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Buyer'
      produces:
      - application/json
      responses:
        "201":
          description: Buyer signed up successfully
          schema:
            type: string
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to sign up
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new buyer
      tags:
      - Buyers
...
swagger: "2.0"